requirements/base.txt
Django==1.8.2
django-bower==5.0.4
django-crispy-forms==1.4.0
djangorestframework==3.1.3
feedparser==5.1.3
Pillow==2.9.0
psycopg2==2.6.1
pydub==0.14.0
python-dateutil==2.4.2
PyYAML==3.11
requests==2.7.0
Unipath==1.1
wheel==0.24.0
